requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
scikit-learn>=1.3,<1.6
xgboost>=1.7,<2.3
pandas>=1.5,<2.3
joblib>=1.3
python-dotenv>=1.0
numpy>=1.24,<2.0
